refactor: requestMappings 일급 컬렉션 분리
